Reworked the entire MatrixFactorization data loading system. It is finally quick enough to work with. Removed RecommendationSystem.Data project as there was no need for it. Decreased accuracy of kNN from double to float for less memory consumption.

+++ RecommendationSystem.Knn/Recommendations/AdjustedWeightedSumRatingAggregator.cs
@@ -2,10 +2,10 @@
 {
     public class AdjustedWeightedSumRatingAggregator : IRatingAggregator
     {
-        public double Aggregate(User user, string artist)
+        public float Aggregate(User user, string artist)
         {
-            double k = 0.0;
-            double r = 0.0;
+            float k = 0.0f;
+            float r = 0.0f;
             foreach (var u in user.Neighbours)
             {
                 k += u.Estimate;
